% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saleOrderTable.R
\name{saleOrderTable_query}
\alias{saleOrderTable_query}
\title{查询销售订单统计表}
\usage{
saleOrderTable_query(
  token,
  FStartDate,
  FEndDate,
  FIsFree,
  Fstatus,
  FBIllType,
  FOrgNumber
)
}
\arguments{
\item{token}{中台数据库token}

\item{FNumber}{单据编号}

\item{FName}{ERP账套名}
}
\value{

}
\description{
查询销售订单统计表
}
\examples{
saleOrderTable_query('B3455A86-2376-4EF3-8000-39AA8B61B759','2023-08-01','2023-08-22','否','全部',"'标准销售订单','样品销售订单'",'102')
}
